+++ packages/ui/src/components/va-modal/Changelog.md
@@ -3,3 +3,5 @@
 - modal sizes moved to sizesConfig
 - added auto size
 - visually increased paddings
+- removed useless `.va-modal__overlay` css variables like: position, left, top, width, height
+- improve z-index handling when multiple modals and dropdowns used